+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## v0.10.0 (2023-10-05)
+* Restructured command / argument interface
+
 ## v0.9.8 (2022-12-06)
 * Fixed list argument modifier